change to contain and not contain
