+++ oap-server/oal-grammar/src/main/antlr4/org/apache/skywalking/oal/rt/grammar/OALLexer.g4
@@ -90,8 +90,8 @@ LESS_EQUAL:                          '<=';
 NOT_EQUAL:                           '!=';
 LIKE:                                'like';
 IN:                                  'in';
-INCLUDES:                            'includes';
-EXCLUDES:                            'excludes';
+CONTAIN:                            'contain';
+NOT_CONTAIN:                        'not contain';
 
 // Literals
